Call _chkstk() on Windows
Call _chkstk() equivalent on Windows when the size of local variables is
greater or equal 4KB. This is needed to properly grow/commit the stack,
page by page, and move the stack guard page. If not done (if the stack
is accessed beyond the guard page), the app will crash.