Floats!
Floating point is supported in 32-bit mode(l)s only.

double is an alias for float. IOW, only 32-bit
single precision floats are supported. This isn't
conformant, but OK for some embedded systems (e.g.
RetroBSD) and simple compilers like this.

Also, the following operators are not supported with
floats at the moment: ++, --, +=, -=, *=, /=.
But +, -, *, /, =, ||, &&, ?:, !, comparison, casts,
if/while/for are OK.

Floating point tests added/adjusted.

This also fixes memcmp().

Still to be done:
- support for floats in *scanf(), *printf()
- difftime()
- float.h
- math.h and its functions
- whatever's missing/forgotten