Fix handling of %TEMP% & %TMP% in tmpnam()/tmpfile()
The TEMP and TMP environment variables in DOS/Windows
were not properly handled in tmpnam()/tmpfile() w.r.t.
slashes in file names resulting in fallback to C:\ as
the location for temporary files.
While in most cases this was OK, this posed a problem
for write-protected C: drives (e.g. USB sticks).