Fix/improve *scanf(); scan floats
This adds support for floating point, yay!
This improves/fixes error handling.
This adds a sscanf() test.

There's a wide range of interpretation of the
C standard (because of poor wording and scarce
examples?) with implementations of *scanf() in
different compilers doing different things,
often clearly violating the standard.

This may need to be revisited.