Straighten PE/COFF import tables in .EXEs
This is to try to reduce the number of false positives
in the various antivirus programs.

This may be a breaking change w.r.t. object/library
files compiled with earlier versions of Smaller C
(especially the ones involved in importing from DLLs).

n2f still needs to be updated to support additional
assembly directives, most notably for alignment.