Split build flags, honour external values
CFLAGS and CC (among others) are commonly set by distribution build
infrastructure. It's helpful if build systems apply those settings
when appropriate; to do this in SmallerC, this changes CFLAGS and CC
to only be set if they don't have an existing value.

C definitions are moved to CPPFLAGS, and appended to any existing
values.

The smrlpp build is changed to apply CPPFLAGS and LDFLAGS as well as
CFLAGS. (This is the default in Make, so it already applies to the
other tools which rely on built-in rules.)

Signed-off-by: Stephen Kitt <steve@sk2.org>