Introduce 80386+ checks at startup
Do this only where a 80386+ is absolutely required,
in the huge and unreal memory mode(l)s (the huge
mode(l) being the default in the compiler when
compiling for DOS).

The DPMI stub normally fails if there's no 32-bit DPMI host,
it just needed the code to be 8086-compatible until the
point where a 80386+ is known to be present.

The tiny and small memory mode(l)s aren't touched now.