Issue #51: extend MZ header to 64 bytes
This lets one append an NE, LE, PE, etc executable to the DOS
executable and set the dword at offset 0x3C of the header
to point to it, thus making a fat executable containing a DOS
part/stub and a part for some other OS/purpoose.

Or one can simply store additional data in the header if desired.

Note, the DPMI stub does not depend on nor utilizes this
mechanism to locate the appended 32-bit protected mode part.